requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# Database
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Twilio
twilio>=8.0.0

# AWS S3/R2 compatible
boto3>=1.34.0

# HTTP Client (for Etherscan)
httpx>=0.26.0

# Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Email validation
email-validator>=2.1.0
